moving entity and smoll fi
