+++ Assets/Scripts/OrchestraArmy/Entity/Controllers/IMovementController.cs
@@ -1,11 +1,13 @@
-﻿namespace OrchestraArmy.Entity.Controllers
+﻿using OrchestraArmy.Entity.Entities;
+
+namespace OrchestraArmy.Entity.Controllers
 {
     public interface IMovementController
     {
         /// <summary>
         /// The entity whose movement is to be controlled.
         /// </summary>
-        public LivingDirectionalEntity Entity { get; set; }
+        public MovingEntity Entity { get; set; }
         
         /// <summary>
         /// Handle the entity's movement.
