+++ Assets/Scripts/OrchestraArmy/Entity/Entities/Projectiles/Controllers/ProjectileMovementController.cs
@@ -5,7 +5,7 @@ namespace OrchestraArmy.Entity.Entities.Projectiles.Controllers
 {
     public class ProjectileMovementController: IMovementController
     {
-        public LivingDirectionalEntity Entity { get; set; }
+        public MovingEntity Entity { get; set; }
         public void HandleMovement()
         {
             var projectile = (Projectile) Entity;
