+++ Assets/Scripts/OrchestraArmy/Entity/LivingDirectionalEntity.cs
@@ -3,28 +3,15 @@ using System.Collections;
 using System.Collections.Generic;
 using OrchestraArmy.Entity.Controllers;
 using OrchestraArmy.Entity.Data;
+using OrchestraArmy.Entity.Entities;
 using UnityEngine;
 
 namespace OrchestraArmy.Entity
 {
-    public abstract class LivingDirectionalEntity : DirectionalEntity
+    public abstract class LivingDirectionalEntity : MovingEntity
     {
-        
-        public Rigidbody RigidBody { get; set; }
-        
         [field: SerializeField]
         public LivingEntityData EntityData { get; set; }
-
-        /// <summary>
-        /// The controller for the entity's movement.
-        /// </summary>
-        public IMovementController MovementController { get; set; }
-
-        protected override void Awake()
-        {
-            base.Awake();
-            this.RigidBody = GetComponent<Rigidbody>();
-        }
     }
    
 }
